requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cmsplugin-filer==1.1.3
Django==1.10.7
django-appconf==1.0.2
django-autoslug==1.9.3
django-chartit==0.2.9
django-chartjs==1.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-crispy-forms==1.7.0
django-dash==0.5.3
django-filer==1.2.8
django-form-utils==1.0.3
django-formtools==2.1
django-mptt==0.8.7
django-nine==0.1.13
django-polymorphic==1.0.2
django-pygments==0.1
django-rest-framework==0.1.0
django-sekizai==0.10.0
django-simple-captcha==0.5.5
django-suit==0.2.25
django-tinymce==2.6.0
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
djangorestframework==3.7.1
easy-thumbnails==2.3
feedparser==5.2.1
html5lib==0.9999999
idna==2.6
lipsum==0.1.2
lorem-ipsum-generator==0.3
MySQL-python==1.2.5
olefile==0.44
ordereddict==1.1
pif==0.8.2
Pillow==4.3.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
six==1.11.0
transliterate==1.10
Unidecode==0.4.21
urllib3==1.22
vishap==0.1.5
